requirements.txt
hdbcli>=2.17.0
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.30.0
plotly>=5.18.0
pyyaml>=6.0
